+++ src/main/java/org/itpc_advanced/local/Text.java
@@ -1,0 +1,6 @@
+package org.itpc_advanced.local;
+
+public class Text {
+	public final String MESSAGE = "+++"; 
+
+}
