+++ src/main/java/org/itpc_advanced/service/NoiseFilter.java
@@ -1,0 +1,5 @@
+package org.itpc_advanced.service;
+
+public class NoiseFilter {
+
+}
